+++ app/src/main/res/layout/recyclerview_contest_item.xml
@@ -3,6 +3,10 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
+    android:layout_marginTop="4dp"
+    android:layout_marginLeft="8dp"
+    android:layout_marginRight="8dp"
+    android:layout_marginBottom="4dp"
     app:cardCornerRadius="2dp"
     app:cardElevation="4dp"
     app:contentPadding="8dp"
